requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
requests==2.31.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
playwright==1.52.0
youtube-transcript-api
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
slowapi==0.1.9
cryptography==42.0.5
